backend/requirements_temp.txt
fastapi>=0.110.0
uvicorn>=0.27.0
pydantic>=2.6.0
jinja2>=3.1.2
pdfkit>=1.0.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
httpx>=0.27.0